removed .ui from widgets and resources
