+++ src/main/kotlin/recources/UIColors.kt
@@ -1,8 +1,8 @@
 package recources
 
 object UIColors {
-    const val UI_PRIMARY = "#e56060" //ROMAN
-    const val UI_SECINDARY = "#3db5ff" //SUMMERSKY
-    const val UI_NEUTRAL = "#ffffff" //WHITE
-    const val UI_NEUTRAL_TEXT = "#000000" //BLACK
+
+    const val ROMAN = "#e56060"
+    const val SUMMERSKY = "#3db5ff"
+
 }
