--- src/main/kotlin/recources/ui/UIColors.kt
@@ -1,8 +1,0 @@
-package recources.ui
-
-object UIColors {
-
-    const val ROMAN = "#e56060"
-    const val SUMMERSKY = "#3db5ff"
-
-}
